split out NinSnesSeq.cpp pitch slide and fade logic + vibrato/tremolo into its own file
